refactor IframeContainer to function component
